# docs/openapi.yaml
openapi: 3.1.0
info:
  title: SkillTown Learning API
  version: 1.0.0
paths:
...
  /api/v1/session:
    post:
      summary: Create Session
      operationId: createSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateSessionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      summary: Delete Session
      operationId: deleteSession
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/town:
    get:
      summary: Get Town
      operationId: getTown
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TownResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/attempts:
    post:
      summary: Create Attempt
      operationId: createAttempt
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAttemptRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttemptResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/attempts/{attempt_id}:
    get:
      summary: Get Attempt
      operationId: getAttempt
      parameters:
      - name: attempt_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Attempt Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttemptResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/attempts/{attempt_id}/respond:
    post:
      summary: Respond
      operationId: respondToAttempt
      parameters:
      - name: attempt_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Attempt Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
              - $ref: '#/components/schemas/ChoiceResponseRequest'
              - $ref: '#/components/schemas/TextResponseRequest'
              discriminator:
                propertyName: kind
                mapping:
                  choice: '#/components/schemas/ChoiceResponseRequest'
                  text: '#/components/schemas/TextResponseRequest'
              title: Body
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttemptResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/attempts/{attempt_id}/hint:
    post:
      summary: Request Hint
      operationId: requestHint
      parameters:
      - name: attempt_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Attempt Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MutationRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HintResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/attempts/{attempt_id}/rewind:
    post:
      summary: Rewind
      operationId: rewindAttempt
      parameters:
      - name: attempt_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Attempt Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MutationRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttemptResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/attempts/{attempt_id}/activity:
    post:
      summary: Record Activity
      operationId: recordActivity
      parameters:
      - name: attempt_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Attempt Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivityRequest'
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/passport:
    get:
      summary: Get Passport
      operationId: getPassport
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PassportResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/recommendations:
    post:
      summary: Recommendations
      operationId: generateRecommendations
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecommendationRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecommendationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    ActivityRequest:
      properties:
        client_event_id:
          type: string
          format: uuid
          title: Client Event Id
        kind:
          type: string
          enum:
          - start
          - heartbeat
          - pause
          - resume
          - end
          title: Kind
      additionalProperties: false
      type: object
      required:
      - client_event_id
      - kind
      title: ActivityRequest
...
    AttemptResponse:
      properties:
        attempt_id:
          type: string
          format: uuid
          title: Attempt Id
        scenario_id:
          type: string
          title: Scenario Id
        scenario_version:
          type: string
          title: Scenario Version
        mode:
          $ref: '#/components/schemas/AttemptMode'
        status:
          $ref: '#/components/schemas/AttemptStatus'
        revision:
          type: integer
          minimum: 0.0
          title: Revision
        assisted:
          type: boolean
          title: Assisted
        node:
          anyOf:
          - $ref: '#/components/schemas/ScenarioNode'
          - type: 'null'
        feedback:
          anyOf:
          - $ref: '#/components/schemas/Feedback'
          - type: 'null'
        effect:
          $ref: '#/components/schemas/Effect'
        learning_updates:
          items:
            $ref: '#/components/schemas/LearningUpdate'
          type: array
          title: Learning Updates
        is_complete:
          type: boolean
          title: Is Complete
        feedback_mode:
          $ref: '#/components/schemas/FeedbackMode'
        assessment_status:
          type: string
          enum:
          - assessed
          - deferred
          - not_requested
          title: Assessment Status
          default: not_requested
        dialogue:
          items:
            $ref: '#/components/schemas/DialogueTurn'
          type: array
          title: Dialogue
        pressure:
          anyOf:
          - $ref: '#/components/schemas/Pressure'
          - type: 'null'
        timing:
          $ref: '#/components/schemas/Timing'
      additionalProperties: false
      type: object
      required:
      - attempt_id
      - scenario_id
      - scenario_version
      - mode
      - status
      - revision
      - assisted
      - node
      - effect
      - is_complete
      - feedback_mode
      - timing
      title: AttemptResponse
...
    ChoiceResponseRequest:
      properties:
        client_event_id:
          type: string
          format: uuid
          title: Client Event Id
        expected_revision:
          type: integer
          minimum: 0.0
          title: Expected Revision
        kind:
          type: string
          const: choice
          title: Kind
        choice_id:
          type: string
          title: Choice Id
      additionalProperties: false
      type: object
      required:
      - client_event_id
      - expected_revision
      - kind
      - choice_id
      title: ChoiceResponseRequest
    CreateAttemptRequest:
      properties:
        scenario_id:
          type: string
          title: Scenario Id
        mode:
          $ref: '#/components/schemas/AttemptMode'
      additionalProperties: false
      type: object
      required:
      - scenario_id
      - mode
      title: CreateAttemptRequest
    CreateSessionRequest:
      properties:
        display_name:
          type: string
          maxLength: 40
          minLength: 1
          title: Display Name
      additionalProperties: false
      type: object
      required:
      - display_name
      title: CreateSessionRequest
    CreateSessionResponse:
      properties:
        session_token:
          type: string
          title: Session Token
        session:
          $ref: '#/components/schemas/SessionView'
      additionalProperties: false
      type: object
      required:
      - session_token
      - session
      title: CreateSessionResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    HintResponse:
      properties:
        attempt_id:
          type: string
          format: uuid
          title: Attempt Id
        revision:
          type: integer
          title: Revision
        assisted:
          type: boolean
          const: true
          title: Assisted
          default: true
        hint:
          type: string
          title: Hint
        policy_card:
          $ref: '#/components/schemas/PolicyCard'
      additionalProperties: false
      type: object
      required:
      - attempt_id
      - revision
      - hint
      - policy_card
      title: HintResponse
...
    MutationRequest:
      properties:
        client_event_id:
          type: string
          format: uuid
          title: Client Event Id
        expected_revision:
          type: integer
          minimum: 0.0
          title: Expected Revision
      additionalProperties: false
      type: object
      required:
      - client_event_id
      - expected_revision
      title: MutationRequest
...
    PassportResponse:
      properties:
        skills:
          items:
            $ref: '#/components/schemas/SkillPassportEntry'
          type: array
          title: Skills
        total_active_seconds:
          type: integer
          minimum: 0.0
          title: Total Active Seconds
        total_model_wait_seconds:
          type: integer
          minimum: 0.0
          title: Total Model Wait Seconds
      additionalProperties: false
      type: object
      required:
      - skills
      - total_active_seconds
      - total_model_wait_seconds
      title: PassportResponse
...
    RecommendationRequest:
      properties:
        max_items:
          type: integer
          maximum: 3.0
          minimum: 1.0
          title: Max Items
          default: 3
      additionalProperties: false
      type: object
      title: RecommendationRequest
    RecommendationResponse:
      properties:
        generated_at:
          type: string
          format: date-time
          title: Generated At
        items:
          items:
            $ref: '#/components/schemas/RecommendationView'
          type: array
          maxItems: 3
          title: Items
      additionalProperties: false
      type: object
      required:
      - generated_at
      - items
      title: RecommendationResponse
...
    TextResponseRequest:
      properties:
        client_event_id:
          type: string
          format: uuid
          title: Client Event Id
        expected_revision:
          type: integer
          minimum: 0.0
          title: Expected Revision
        kind:
          type: string
          const: text
          title: Kind
        text:
          type: string
          maxLength: 1000
          minLength: 1
          title: Text
      additionalProperties: false
      type: object
      required:
      - client_event_id
      - expected_revision
      - kind
      - text
      title: TextResponseRequest
...
    TownResponse:
      properties:
        categories:
          items:
            $ref: '#/components/schemas/CategoryView'
          type: array
          title: Categories
        npcs:
          items:
            $ref: '#/components/schemas/TownNpc'
          type: array
          title: Npcs
        screening:
          anyOf:
          - $ref: '#/components/schemas/TaskSummary'
          - type: 'null'
      additionalProperties: false
      type: object
      required:
      - categories
      - npcs
      title: TownResponse
...